Fixed failes to return new employee id
Fails becuase its not possible to cast boxed UInt64 directly to int

+++ Assignment-3/BookingSystem/Storage/CustomerStorage.cs
@@ -66,7 +66,7 @@ namespace BookingSystem
             command.Parameters.AddWithValue("@birthdate", customer.Birthdate);
 
             conn.Open();
-            return (int)command.ExecuteScalar();
+            return (int)(ulong)command.ExecuteScalar();
         }
     }
 }
